* commit 'f4639fd266c49881faf5d6822d7dd6bd926cf807': (33 commits)
  resolved ut
  change color qfb
  grafica
  finish all
  Actualizacion de chart
  Validación de materia prima requerida.
  Se corrige logs
  Validación de materia prima requerida.
  Actualizacion de productividad
  Actualizacion de grafica
  init add client to find orders
  Se agrega fix para status manual
  versión revisión build 16
  Se corrige error de busqueda para En proceso
  versión revisión build 15
  fix lotes asignados de servidor
  ajustes a pantalla detalle de orden
  Se corrige error en fechas
  Se genera versión para revision de issues 560, 549, 477, 566
  Se realiza la funcionalidad para los comentarios en lotes
  ...

+++ services/sapadapter-service/Omicron.SapAdapter.DataAccess/DAO/Sap/SapDao.cs
@@ -261,7 +261,7 @@ namespace Omicron.SapAdapter.DataAccess.DAO.Sap
                                     Description = p.LargeDescription,
                                     BaseQuantity = w.BaseQuantity,
                                     RequiredQuantity = w.RequiredQty,
-                                    Consumed = (int)w.ConsumidoQty,
+                                    Consumed = w.ConsumidoQty,
                                     Available = dp.OnHand - dp.IsCommited + dp.OnOrder,
                                     Unit = w.UnidadCode,
                                     Warehouse = w.Almacen,
